@@ -7,7 +7,7 @@ class HomeHeaderView extends Component {
     return (
       <div className="HomeHeaderView">
         <img src="/img/bars-solid.svg" alt="Menu" className="hamburger" onClick={this.props.onHamburgerClick}/>
-        <img className="logo" src="/img/dormspam.svg" alt="Logo" />
+        <img className="logo" src="/img/dormdigest.svg" alt="Logo" />
         <Link to="/settings">
           {/*<img src="/img/settings.svg" alt="Settings" />*/}
         </Link>
