@@ -8,20 +8,20 @@ class HomeSidebarEventModal extends Component {
     super(props);
 
     this.state = {
-      colors: {}
+      colors: {...Categories.getCategoriesColorMapping()}  //Shallow copy
     };
 
     this.getCategories = this.getCategories.bind(this);
     
-    Categories.getCategories().then(response => {
-      for (let i = 0; i < response.data.length; i++) {
-        this.state.colors[response.data[i].name] = response.data[i]["color"];
-      }
-    });
+    // Categories.getCategoriesColorMapping().then(response => {
+    //   for (let i = 0; i < response.data.length; i++) {
+    //     this.state.colors[response.data[i].name] = response.data[i]["color"];
+    //   }
+    // });
   }
 
   getCategories() {
-    let categories = this.props.event.categories;
+    let categories = this.props.event.tags;
     let tags = [];
     for (let i=0; i < categories.length; i++) {
       tags.push(<span className="tags"
@@ -40,7 +40,7 @@ class HomeSidebarEventModal extends Component {
     }
 
     let categoryTags = this.getCategories();
-    let date = this.props.event.sent_from.split("on")[0] + " on " + moment(new Date(this.props.event.sent_from.split("on")[1]).toString()).format('MM/DD/YYYY h:mm a')
+    let date = this.props.event.user_email + " on " + moment(new Date(this.props.event.date_created).toString()).format('MM/DD/YYYY h:mm a')
 
     return (
       <div className="HomeSidebarEventModal">
@@ -52,9 +52,10 @@ class HomeSidebarEventModal extends Component {
             {categoryTags}
           </h3>
           <div className="padding"></div>
+          <h2>{this.props.event.title}</h2>
           <hr />
         </div>
-        <div dangerouslySetInnerHTML={{__html: this.props.event.description}} />
+        <div dangerouslySetInnerHTML={{__html: this.props.event.description_html}} />
       </div>
     );
   }
